requirements.txt
aiogoogletrans==3.3.2
aiohttp==3.6.3
aioredis==1.3.1
aiosqlite==0.16.1
aiozaneapi==1.6
akinator.py==4.1.0
altgraph==0.17
appdirs==1.4.4
arrow==0.17.0
astunparse==1.6.3
async-cleverbot==0.2.2
async-cse==0.3.0
async-rediscache==0.2.0
async-timeout==3.0.1
asyncdagpi==3.2.2
asyncpg==0.22.0
asyncpraw==7.1.1
asyncprawcore==1.5.1
attrs==20.3.0
autopep8==1.5.4
beautifulsoup4==4.9.3
better-profanity==0.7.0
braceexpand==0.1.6
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
chatbotAI==0.3.0.0
cleverbotfree==1.2.7
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@6f2b873bf0d28903eb752aa1166b3aac26dc9007
discord-ext-owoify==0.0.5a0
discord-py-slash-command==1.0.9.4
discord-webhook==0.11.0
discord.py==1.6.0
distlib==0.3.1
dontimport==1.0.0
emoji==1.2.0
essential-generators==1.0
filelock==3.0.12
fuzzywuzzy==0.18.0
google-search==1.0.2
google-trans-new==1.1.9
googletrans==3.1.0a0
gtrending==0.3.1
gTTS==2.2.2
h11==0.9.0
h2==3.2.0
hatesonar==0.0.7
hewwwo==1.0.6
hiredis==1.1.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
humanize==3.2.0
hyperframe==5.2.0
idna==2.10
import-expression==1.1.4
jishaku==1.20.0.220
joblib==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
macholib==1.14
matplotlib==3.3.3
multidict==4.7.6
notify-run==0.0.13
numpy==1.20.1
owoify-py==1.0.5
owotext==1.1.1
pandas==1.2.1
Pillow==8.1.0
pipenv==2020.11.15
polaroid==0.3.1
psutil==5.8.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.4
pyinstaller-hooks-contrib==2020.11
PyNaCl==1.4.0
pync==2.0.3
pynput==1.7.2
pyobjc-core==7.1
pyobjc-framework-Cocoa==7.1
pyobjc-framework-Quartz==7.1
pyparsing==2.4.7
PyQRCode==1.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-Levenshtein==0.12.2
pytz==2021.1
random-cat==1.0.1
requests==2.25.1
rfc3986==1.4.0
rich==7.1.0
scikit-learn==0.24.1
scipy==1.6.0
selenium==3.141.0
sentry-sdk==0.19.5
similar==0.2
simpleeval==0.9.10
six==1.15.0
sniffio==1.2.0
soupsieve==2.2
sr-api==1.1.3
starcli==2.16.0
threadpoolctl==2.1.0
toml==0.10.2
twemoji-parser==0.4.3
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.3
vacefron.py==1.6.1
virtualenv==20.4.2
virtualenv-clone==0.5.4
wavelink==0.9.6
wtfpython==0.2.1
yarl==1.5.1
youtube-dl==2021.2.10
